+++ test/sql/test.sql
@@ -99,3 +99,12 @@ select pg_ddl_script('uuid'::regtype);
 select pg_ddl_script('text'::regtype);
 select pg_ddl_script('xml'::regtype);
 
+create view test_class_v_opt1 with (security_barrier) 
+as select * from test_class_v order by 1;
+create view test_class_v_opt2 
+as select * from test_class_v order by 1;
+alter  view test_class_v_opt2 set (security_barrier='true');
+
+select pg_ddl_script('test_class_v_opt1'::regclass);
+select pg_ddl_script('test_class_v_opt2'::regclass);
+